fix bug integrate 1

+++ Final_Aptech_Sem4_Movie/src/main/java/com/demo/services/admin/PriceService.java
@@ -20,7 +20,7 @@ public interface PriceService {
 	
 	public boolean create (Price price);
 	
-	public Price findById (int id);
+	public Price findById (long id);
 	
 	public boolean delete (int id);
 }
